requirements.txt
pytest
cupy-cuda12x
scipy
Pillow